Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicyviolation-base-uri-deny-expected.txt

CONSOLE MESSAGE: Refused to set the document's base URI to 'http://example.com/' because it violates the following Content Security Policy directive: "base-uri 'self'".

Check that base URIs cannot be set if they violate the page's policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Kicking off the tests:
PASS document.baseURI is document.location.href
PASS window.e.documentURI is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/base-uri-deny.html"
PASS window.e.referrer is ""
PASS window.e.blockedURI is "http://example.com"
PASS window.e.violatedDirective is "base-uri 'self'"
PASS window.e.effectiveDirective is "base-uri"
PASS window.e.originalPolicy is "base-uri 'self'"
PASS window.e.sourceFile is "http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/base-uri-deny.html"
PASS window.e.lineNumber is 24
PASS successfullyParsed is true

TEST COMPLETE
